+++ src/main/kotlin/no/nav/amt/person/service/clients/pdl/PdlClient.kt
@@ -1,16 +1,8 @@
 package no.nav.amt.person.service.clients.pdl
 
-import no.nav.amt.person.service.person.model.Adresse
 import no.nav.amt.person.service.person.model.AdressebeskyttelseGradering
-import no.nav.amt.person.service.person.model.Bostedsadresse
 import no.nav.amt.person.service.person.model.IdentType
-import no.nav.amt.person.service.person.model.Kontaktadresse
-import no.nav.amt.person.service.person.model.Matrikkeladresse
-import no.nav.amt.person.service.person.model.Oppholdsadresse
 import no.nav.amt.person.service.person.model.Personident
-import no.nav.amt.person.service.person.model.Postboksadresse
-import no.nav.amt.person.service.person.model.Vegadresse
-import no.nav.amt.person.service.poststed.Postnummer
 import no.nav.amt.person.service.poststed.PoststedRepository
 import no.nav.amt.person.service.utils.GraphqlUtils
 import no.nav.amt.person.service.utils.GraphqlUtils.GraphqlResponse
@@ -20,6 +12,7 @@ import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.RequestBody.Companion.toRequestBody
 import org.slf4j.LoggerFactory
+import org.springframework.http.HttpHeaders
 import tools.jackson.databind.ObjectMapper
 import tools.jackson.module.kotlin.readValue
 import java.util.function.Supplier
@@ -66,7 +59,13 @@ class PdlClient(
 				throw RuntimeException("PDL respons inneholder ikke data")
 			}
 
-			return toPdlBruker(gqlResponse.data)
+			val pdlPerson = gqlResponse.data.toPdlBruker { postnummer -> poststedRepository.getPoststeder(postnummer) }
+
+			if (pdlPerson.etternavn == UNKNOWN_NAME) {
+				log.warn("PDL-person inneholder navn: $UNKNOWN_NAME for ident $personident")
+			}
+
+			return gqlResponse.data.toPdlBruker { postnummer -> poststedRepository.getPoststeder(postnummer) }
 		}
 	}
 
@@ -167,7 +166,8 @@ class PdlClient(
 				throw RuntimeException("PDL respons inneholder ikke data")
 			}
 
-			return getTelefonnummer(gqlResponse.data.hentPerson.telefonnummer)
+			return gqlResponse.data.hentPerson.telefonnummer
+				.toTelefonnummer()
 		}
 	}
 
@@ -196,7 +196,8 @@ class PdlClient(
 				throw RuntimeException("PDL respons inneholder ikke data")
 			}
 
-			return getDiskresjonskode(gqlResponse.data.hentPerson.adressebeskyttelse)
+			return gqlResponse.data.hentPerson.adressebeskyttelse
+				.toDiskresjonskode()
 		}
 	}
 
@@ -204,173 +205,11 @@ class PdlClient(
 		Request
 			.Builder()
 			.url("$baseUrl/graphql")
-			.addHeader("Authorization", "Bearer ${tokenProvider.get()}")
+			.addHeader(HttpHeaders.AUTHORIZATION, "Bearer ${tokenProvider.get()}")
 			.addHeader("Tema", "GEN")
 			.addHeader("behandlingsnummer", BEHANDLINGSNUMMER)
 			.post(jsonPayload.toRequestBody(mediaTypeJson))
 			.build()
-
-	private fun toPdlBruker(response: PdlQueries.HentPerson.ResponseData): PdlPerson {
-		val navn = response.hentPerson.navn.firstOrNull() ?: throw RuntimeException("PDL person mangler navn")
-		val telefonnummer = getTelefonnummer(response.hentPerson.telefonnummer)
-		val diskresjonskode = getDiskresjonskode(response.hentPerson.adressebeskyttelse)
-
-		return PdlPerson(
-			fornavn = navn.fornavn,
-			mellomnavn = navn.mellomnavn,
-			etternavn = navn.etternavn,
-			telefonnummer = telefonnummer,
-			adressebeskyttelseGradering = diskresjonskode,
-			identer =
-				response.hentIdenter.identer.map {
-					Personident(
-						it.ident,
-						it.historisk,
-						IdentType.valueOf(it.gruppe),
-					)
-				},
-			adresse = getAdresse(response.hentPerson),
-		)
-	}
-
-	private fun getTelefonnummer(telefonnummere: List<PdlQueries.Attribute.Telefonnummer>): String? {
-		val prioritertNummer = telefonnummere.minByOrNull { it.prioritet } ?: return null
-
-		return "${prioritertNummer.landskode}${prioritertNummer.nummer}"
-	}
-
-	private fun getDiskresjonskode(adressebeskyttelse: List<PdlQueries.Attribute.Adressebeskyttelse>): AdressebeskyttelseGradering? =
-		when (adressebeskyttelse.firstOrNull()?.gradering) {
-			"STRENGT_FORTROLIG_UTLAND" -> AdressebeskyttelseGradering.STRENGT_FORTROLIG_UTLAND
-			"STRENGT_FORTROLIG" -> AdressebeskyttelseGradering.STRENGT_FORTROLIG
-			"FORTROLIG" -> AdressebeskyttelseGradering.FORTROLIG
-			"UGRADERT" -> AdressebeskyttelseGradering.UGRADERT
-			else -> null
-		}
-
-	private fun getAdresse(hentPersonResponse: PdlQueries.HentPerson.HentPerson): Adresse? {
-		val kontaktadresseFraPdl = hentPersonResponse.kontaktadresse.firstOrNull()
-		val bostedsadresseFraPdl = hentPersonResponse.bostedsadresse.firstOrNull()
-		val oppholdsadresseFraPdl = hentPersonResponse.oppholdsadresse.firstOrNull()
-
-		val unikePostnummer =
-			listOfNotNull(
-				kontaktadresseFraPdl?.vegadresse?.postnummer,
-				kontaktadresseFraPdl?.postboksadresse?.postnummer,
-				bostedsadresseFraPdl?.vegadresse?.postnummer,
-				bostedsadresseFraPdl?.matrikkeladresse?.postnummer,
-				oppholdsadresseFraPdl?.vegadresse?.postnummer,
-				oppholdsadresseFraPdl?.matrikkeladresse?.postnummer,
-			).distinct()
-
-		val poststeder = poststedRepository.getPoststeder(unikePostnummer)
-
-		if (poststeder.isEmpty()) {
-			return null
-		}
-
-		val adresse =
-			Adresse(
-				bostedsadresse = bostedsadresseFraPdl?.toBostedsadresse(poststeder),
-				oppholdsadresse = oppholdsadresseFraPdl?.toOppholdsadresse(poststeder),
-				kontaktadresse = kontaktadresseFraPdl?.toKontaktadresse(poststeder),
-			)
-
-		if (adresse.bostedsadresse == null && adresse.oppholdsadresse == null && adresse.kontaktadresse == null) {
-			return null
-		}
-		return adresse
-	}
-
-	private fun PdlQueries.Attribute.Bostedsadresse.toBostedsadresse(poststeder: List<Postnummer>): Bostedsadresse? {
-		if (vegadresse == null && matrikkeladresse == null) {
-			return null
-		}
-		val bostedsadresse =
-			Bostedsadresse(
-				coAdressenavn = coAdressenavn,
-				vegadresse = vegadresse?.toVegadresse(poststeder),
-				matrikkeladresse = matrikkeladresse?.toMatrikkeladresse(poststeder),
-			)
-		if (bostedsadresse.vegadresse == null && bostedsadresse.matrikkeladresse == null) {
-			return null
-		}
-		return bostedsadresse
-	}
-
-	private fun PdlQueries.Attribute.Oppholdsadresse.toOppholdsadresse(poststeder: List<Postnummer>): Oppholdsadresse? {
-		if (vegadresse == null && matrikkeladresse == null) {
-			return null
-		}
-		val oppholdsadresse =
-			Oppholdsadresse(
-				coAdressenavn = coAdressenavn,
-				vegadresse = vegadresse?.toVegadresse(poststeder),
-				matrikkeladresse = matrikkeladresse?.toMatrikkeladresse(poststeder),
-			)
-		if (oppholdsadresse.vegadresse == null && oppholdsadresse.matrikkeladresse == null) {
-			return null
-		}
-		return oppholdsadresse
-	}
-
-	private fun PdlQueries.Attribute.Kontaktadresse.toKontaktadresse(poststeder: List<Postnummer>): Kontaktadresse? {
-		if (vegadresse == null && postboksadresse == null) {
-			return null
-		}
-		val kontaktadresse =
-			Kontaktadresse(
-				coAdressenavn = coAdressenavn,
-				vegadresse = vegadresse?.toVegadresse(poststeder),
-				postboksadresse = postboksadresse?.toPostboksadresse(poststeder),
-			)
-		if (kontaktadresse.vegadresse == null && kontaktadresse.postboksadresse == null) {
-			return null
-		}
-		return kontaktadresse
-	}
-
-	private fun PdlQueries.Attribute.Vegadresse.toVegadresse(poststeder: List<Postnummer>): Vegadresse? {
-		if (postnummer == null) {
-			return null
-		}
-		val poststed = poststeder.find { it.postnummer == postnummer } ?: return null
-
-		return Vegadresse(
-			husnummer = husnummer,
-			husbokstav = husbokstav,
-			adressenavn = adressenavn,
-			tilleggsnavn = tilleggsnavn,
-			postnummer = postnummer,
-			poststed = poststed.poststed,
-		)
-	}
-
-	private fun PdlQueries.Attribute.Matrikkeladresse.toMatrikkeladresse(poststeder: List<Postnummer>): Matrikkeladresse? {
-		if (postnummer == null) {
-			return null
-		}
-		val poststed = poststeder.find { it.postnummer == postnummer } ?: return null
-
-		return Matrikkeladresse(
-			tilleggsnavn = tilleggsnavn,
-			postnummer = postnummer,
-			poststed = poststed.poststed,
-		)
-	}
-
-	private fun PdlQueries.Attribute.Postboksadresse.toPostboksadresse(poststeder: List<Postnummer>): Postboksadresse? {
-		if (postnummer == null) {
-			return null
-		}
-		val poststed = poststeder.find { it.postnummer == postnummer } ?: return null
-
-		return Postboksadresse(
-			postboks = postboks,
-			postnummer = postnummer,
-			poststed = poststed.poststed,
-		)
-	}
 
 	private fun throwPdlApiErrors(response: GraphqlResponse<*, PdlQueries.PdlErrorExtension>) {
 		var melding = "Feilmeldinger i respons fra pdl:\n"
