+++ src/main/kotlin/no/nav/amt/person/service/clients/pdl/PdlPerson.kt
@@ -14,6 +14,7 @@ data class PdlPerson(
 	val adressebeskyttelseGradering: AdressebeskyttelseGradering?,
 	val identer: List<Personident>,
 	val adresse: Adresse?,
+	val falskIdentitet: Boolean,
 ) {
 	fun getAdressebeskyttelse(): Adressebeskyttelse? =
 		if (adressebeskyttelseGradering?.erBeskyttet() == true) {
