+++ src/test/kotlin/no/nav/amt/person/service/data/TestData.kt
@@ -121,6 +121,7 @@ object TestData {
 		adressebeskyttelseGradering: AdressebeskyttelseGradering? = null,
 		identer: List<Personident> = listOf(Personident(person.personident, false, IdentType.FOLKEREGISTERIDENT)),
 		adresse: Adresse? = lagAdresse(),
+		falskIdentitet: Boolean = false,
 	) = PdlPerson(
 		fornavn = person.fornavn,
 		mellomnavn = person.mellomnavn,
@@ -129,6 +130,7 @@ object TestData {
 		adressebeskyttelseGradering = adressebeskyttelseGradering,
 		identer = identer,
 		adresse = adresse,
+		falskIdentitet = falskIdentitet,
 	)
 
 	fun lagPersonident(
