+++ src/test/kotlin/no/nav/amt/person/service/person/PersonServiceTest.kt
@@ -49,6 +49,7 @@ class PersonServiceTest {
 						Personident(ident = TestData.randomIdent(), historisk = true, type = identType),
 					),
 				adresse = null,
+				falskIdentitet = false,
 			)
 
 		every { pdlClient.hentPerson(personident) } returns pdlPerson
